Transfer out TeamGroup imp from team-page to component
